requirements.txt
google-genai>=1.0.0
pydantic>=2.0
python-telegram-bot>=22.0
python-jobspy>=1.1
python-dotenv>=1.0
langdetect>=1.0.9
